Facility Speed Type Change

+++ API/API/Models/FacilityModel.cs
@@ -43,7 +43,7 @@ namespace API.Models
         public decimal BreakdownFee { get; set; }
 
         public int Station { get; set; }
-        public int Speed { get; set; }
+        public float Speed { get; set; }
         public int Capacity { get; set; }
         public DateTime? WarrantyFinishDate { get; set; }
         public decimal OldMaintenanceFee { get; set; }
